feat: implement "modules" task
